Le changement c'st maintenant ---> Clean Code

+++ Printf/srcs/basic/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: juspende <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/06 14:25:53 by juspende          #+#    #+#             */
-/*   Updated: 2017/11/29 10:35:41 by juspende         ###   ########.fr       */
+/*   Updated: 2017/11/29 11:15:23 by juspende         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,10 +15,10 @@
 size_t		ft_strlen(char *str)
 {
 	size_t		result;
-	uint32_t	*i;
-	
+	uint64_t	*i;
+
 	result = 0;
-	i = (uint32_t *)str;
+	i = (uint64_t *)str;
 	while (42)
 	{
 		if (!(*i & 0x000000ff))
